OBG-549. Added API security module
