+++ opba-api-security/src/main/java/de/adorsys/opba/api/security/EnableTokenBasedApiSecurity.java
@@ -1,0 +1,20 @@
+package de.adorsys.opba.api.security;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Inherited;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Configuration
+@EnableConfigurationProperties
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@Inherited
+@ComponentScan(basePackageClasses = EnableTokenBasedApiSecurity.class)
+public @interface EnableTokenBasedApiSecurity {
+}
