+++ opba-api-security/src/test/java/de/adorsys/opba/api/security/service/TokenBasedAuthServiceTest.java
@@ -1,0 +1,5 @@
+package de.adorsys.opba.api.security.service;
+
+class TokenBasedAuthServiceTest {
+
+}
