+++ opba-banking-protocol-facade/src/main/java/de/adorsys/opba/protocol/facade/services/password/FintechUserPasswordGenerator.java
@@ -1,6 +1,5 @@
 package de.adorsys.opba.protocol.facade.services.password;
 
-import com.google.crypto.tink.subtle.Base64;
 import de.adorsys.opba.protocol.facade.config.auth.FacadeAuthConfig;
 import de.adorsys.opba.protocol.facade.config.auth.PasswordGenRandomConfig;
 import lombok.RequiredArgsConstructor;
@@ -8,6 +7,7 @@ import lombok.SneakyThrows;
 import org.springframework.stereotype.Service;
 
 import java.security.SecureRandom;
+import java.util.Base64;
 
 @Service
 @RequiredArgsConstructor
@@ -22,6 +22,6 @@ public class FintechUserPasswordGenerator {
         SecureRandom random = passwordGenRandom.getRandom();
         byte[] bytes = new byte[facadeAuthConfig.getRedirect().getPassword().getByteSize()];
         random.nextBytes(bytes);
-        return Base64.urlSafeEncode(bytes);
+        return Base64.getUrlEncoder().encodeToString(bytes);
     }
 }
