+++ opba-banking-protocol-facade/src/main/java/de/adorsys/opba/protocol/facade/services/psu/PsuAuthService.java
@@ -1,15 +1,10 @@
 package de.adorsys.opba.protocol.facade.services.psu;
 
-import com.nimbusds.jose.JWSHeader;
-import com.nimbusds.jose.JWSSigner;
-import com.nimbusds.jwt.JWTClaimsSet;
-import com.nimbusds.jwt.SignedJWT;
 import de.adorsys.datasafe.encrypiton.api.types.UserIDAuth;
 import de.adorsys.datasafe.types.api.actions.ReadRequest;
 import de.adorsys.datasafe.types.api.actions.WriteRequest;
 import de.adorsys.opba.db.domain.entity.psu.Psu;
 import de.adorsys.opba.db.repository.jpa.psu.PsuRepository;
-import de.adorsys.opba.protocol.facade.config.auth.TppTokenProperties;
 import de.adorsys.opba.protocol.facade.config.encryption.impl.psu.PsuSecureStorage;
 import de.adorsys.opba.protocol.facade.exceptions.PsuAuthenticationException;
 import de.adorsys.opba.protocol.facade.exceptions.PsuAuthorizationException;
@@ -21,10 +16,6 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.time.Duration;
-import java.time.ZoneOffset;
-import java.time.ZonedDateTime;
-import java.util.Date;
 import java.util.Optional;
 
 @Service
@@ -34,9 +25,6 @@ public class PsuAuthService {
     public static final String AUTHENTICATION_CHECKER = "63f510c1-8315-4b7e-8f24-0a2e6a135ea4";
     private final PsuRepository psuRepository;
     private final PsuSecureStorage psuSecureStorage;
-    private final JWSHeader jwsHeader;
-    private final JWSSigner jwsSigner;
-    private final TppTokenProperties tppTokenProperties;
 
     @Transactional(readOnly = true)
     public Psu tryAuthenticateUser(String login, String password) throws PsuAuthorizationException {
@@ -60,20 +48,6 @@ public class PsuAuthService {
         psuSecureStorage.registerPsu(newPsu, password::toCharArray);
         authenticateInDatasafe(newPsu.getUserIdAuth(password::toCharArray));
         return newPsu;
-    }
-
-    @SneakyThrows
-    public String generateToken(String id) {
-        ZonedDateTime currentTime = ZonedDateTime.now(ZoneOffset.UTC);
-        Duration duration = tppTokenProperties.getKeyValidityDuration();
-        JWTClaimsSet claims = new JWTClaimsSet.Builder()
-                .expirationTime(Date.from(currentTime.plus(duration).toInstant()))
-                .issueTime(Date.from(currentTime.toInstant()))
-                .subject(String.valueOf(id))
-                .build();
-        SignedJWT signedJWT = new SignedJWT(jwsHeader, claims);
-        signedJWT.sign(jwsSigner);
-        return signedJWT.serialize();
     }
 
     private void authenticateInDatasafe(UserIDAuth idAuth) throws IOException {
